[CXF-7069] OAuthJPAProvider hibernate optimization
httpVerbs and uris OAuthPermission attributes are cached
in 2nd level cache.

This is achieved by using a specific hbm.xml for Permission
and OAuthPermission.

This hbm.xml replaces JPA definition for these 2 entities
(we cannot override just a part of a JPA entity definition
with hbm.xml, we need to redefine all the mapping).